refactor(encoding): extract level config to levels/config.rs
Move the per-level tuning config out of match_generator.rs into a dedicated
levels/config.rs: the HcConfig / RowConfig / DfastConfig / FastConfig knobs,
LevelParams + LEVEL_TABLE, the public-parameter overrides, the source-size
tiering, and the workspace estimators.

Pure mechanical move with no behaviour change: the code is relocated verbatim
(only the relative `super::` paths rewritten to absolute `crate::encoding::`
for the deeper module, and the resolution API widened to pub(crate)), so the
compressed output stays byte-identical. match_generator.rs drops 1119 lines
(13125 -> 12006). Consumers (frame_compressor, dfast, row) and the public
re-export of the workspace estimators now source the API from levels::config.

Part of #441.
