+++ zstd/src/encoding/levels/mod.rs
@@ -1,2 +1,3 @@
+pub(crate) mod config;
 mod fastest;
 pub(crate) use fastest::{compress_block_encoded, compress_block_encoded_borrowed};
